backend/requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
databases[postgresql]==0.8.0

# 数据源
tushare==1.2.89
akshare==1.12.79
pandas==2.1.4
numpy==1.26.2
ta-lib==0.4.28

# AI集成
openai==1.3.5
zhipuai==2.0.1
deepseek-api==0.1.0
langchain==0.0.350

# 缓存
redis==5.0.1
aioredis==2.0.1

# 工具
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.0

# 监控
sentry-sdk==1.38.0
prometheus-client==0.19.0